+++ src/N26ToYNAB/Program.cs
@@ -8,6 +8,7 @@ using CommandLine;
 using CsvHelper;
 using CsvHelper.Configuration;
 using N26ToYNAB.Model;
+using Serilog;
 
 namespace N26ToYNAB
 {
@@ -16,15 +17,19 @@ namespace N26ToYNAB
         /// <summary>
         /// General CSV parser configuration
         /// </summary>
-        private static readonly CsvConfiguration _csvConfig = new CsvHelper.Configuration.CsvConfiguration(CultureInfo.InvariantCulture)
+        private static readonly CsvConfiguration CsvConfig = new(CultureInfo.InvariantCulture)
         {
             Delimiter = ",",
             Quote = '"',
             HasHeaderRecord = true,
         };
 
-        static void Main(string[] args)
+        private static void Main(string[] args)
         {
+            Log.Logger = new LoggerConfiguration()
+                .WriteTo.Console()
+                .CreateLogger();
+            
             Parser.Default.ParseArguments<Options>(args)
                 .WithParsed(ConversionHandler)
                 .WithNotParsed(ErrorHandler);
@@ -32,7 +37,7 @@ namespace N26ToYNAB
 
         private static void ErrorHandler(IEnumerable<Error> errors)
         {
-            Console.Error.WriteLine("Could not run conversion due to missing input");
+            Log.Logger.Error("Could not run conversion due to missing input");
         }
 
         /// <summary>
@@ -41,24 +46,24 @@ namespace N26ToYNAB
         /// <param name="options">Parsed commandline options</param>
         private static void ConversionHandler(Options options)
         {
-            Console.WriteLine("Started Conversion");
+            Log.Logger.Information("Started Conversion");
             var n26Transactions = ParseData(options.InputCSV);
             var ynabTransaction = ConvertModels(n26Transactions);
-            CreateYnabCSV(ynabTransaction, options.OutputCSV, options.InputCSV);
-            Console.WriteLine("Finished Conversion");
+            CreateYnabCsv(ynabTransaction, options.OutputCSV, options.InputCSV);
+            Log.Logger.Information("Finished Conversion");
         }
 
-        private static void CreateYnabCSV(IEnumerable<YNABTransaction> ynabTransaction, string outputCsvPath,
-            string inputCSVPath)
+        private static void CreateYnabCsv(IEnumerable<YNABTransaction> ynabTransaction, string outputCsvPath,
+            string inputCsvPath)
         {
             if (string.IsNullOrWhiteSpace(outputCsvPath))
             {
-                outputCsvPath = CreateOutputFilePath(inputCSVPath, Config.YNABSuffix);
-                Console.WriteLine($"No output file specified, using: {outputCsvPath}");
+                outputCsvPath = CreateOutputFilePath(inputCsvPath, Config.YNABSuffix);
+                Log.Logger.Information("No output file specified, using: {FileName}", outputCsvPath);
             }
-            
+
             using var writer = new StreamWriter(outputCsvPath);
-            using var csv = new CsvWriter(writer, _csvConfig);
+            using var csv = new CsvWriter(writer, CsvConfig);
             csv.WriteRecords(ynabTransaction);
         }
 
@@ -69,69 +74,62 @@ namespace N26ToYNAB
         /// <returns>List of <see cref="YNABTransaction"/></returns>
         private static IEnumerable<YNABTransaction> ConvertModels(IEnumerable<N26Transaction> n26Transactions)
         {
-            var ynabRecords = new List<YNABTransaction>();
-
-            foreach (var transaction in n26Transactions)
+            var ynabRecords = n26Transactions.Select(transaction => new YNABTransaction
             {
-                ynabRecords.Add(new YNABTransaction()
-                {
-                    Amount = transaction.Amount.ToString(),
-                    Payee = transaction.Recipient,
-                    Date = transaction.Date.ToString("yyyy-MM-dd"),
-                    Memo = CreateYnabMemo(transaction)
-                });
-            }
+                Amount = transaction.Amount.ToString(CultureInfo.InvariantCulture), Payee = transaction.Recipient,
+                Date = transaction.Date.ToString("yyyy-MM-dd"), Memo = CreateYnabMemo(transaction)
+            }).ToList();
 
-            Console.WriteLine($"Converting to YNAB transactions");
+            Log.Logger.Information("Converting to YNAB transactions");
             return ynabRecords;
         }
 
         /// <summary>
         /// Parses n26 csv to model
         /// </summary>
-        /// <param name="inputCSV">Path to n26 csv file</param>
-        /// <returns>List of <see cref="N26Transaction"/> objectss</returns>
-        private static IEnumerable<N26Transaction> ParseData(string inputCSV)
+        /// <param name="inputCsv">Path to n26 csv file</param>
+        /// <returns>List of <see cref="N26Transaction"/> objects</returns>
+        private static IEnumerable<N26Transaction> ParseData(string inputCsv)
         {
-            if (!File.Exists(inputCSV))
+            if (!File.Exists(inputCsv))
             {
                 Console.Error.WriteLine("Could not find input file");
                 Environment.Exit(-1);
             }
 
             List<N26Transaction> records = null;
-            
-            Console.WriteLine($"Trying to parse input file {inputCSV}");
+
+            Log.Logger.Information("Trying to parse input file {File}", inputCsv);
             try
             {
-                using var reader = new StreamReader(inputCSV);
-                using var csv = new CsvReader(reader, _csvConfig);
+                using var reader = new StreamReader(inputCsv);
+                using var csv = new CsvReader(reader, CsvConfig);
                 records = csv.GetRecords<N26Transaction>().ToList();
             }
             catch (Exception e)
             {
-                Console.Error.WriteLine($"Could read / parse input file!{Environment.NewLine}Exception: {e.Message}");
+                Log.Logger.Error(e, "Could read / parse input file!");
                 Environment.Exit(-1);
             }
 
-            Console.WriteLine($"Found {records.Count} N26 transactions");
+            Log.Logger.Information("Found {RecordCount} N26 transactions", records.Count);
             return records;
         }
 
         /// <summary>
         /// Creates output file name by using input file name and appending suffix
         /// </summary>
-        /// <param name="inputCSVPath">input file name</param>
+        /// <param name="inputCsvPath">input file name</param>
         /// <param name="suffix">suffix for output file</param>
         /// <returns>Output filename "INPUT_SUFFIX.csv</returns>
         /// <exception cref="ArgumentException">Thrown if input file name is null or whitespace</exception>
-        private static string CreateOutputFilePath(string inputCSVPath, string suffix)
+        private static string CreateOutputFilePath(string inputCsvPath, string suffix)
         {
-            if (string.IsNullOrWhiteSpace(inputCSVPath))
+            if (string.IsNullOrWhiteSpace(inputCsvPath))
                 throw new ArgumentException("Path cannot be null or whitespace!");
 
-            var folderPath = Path.GetDirectoryName(inputCSVPath);
-            var fileName = Path.GetFileNameWithoutExtension(inputCSVPath);
+            var folderPath = Path.GetDirectoryName(inputCsvPath);
+            var fileName = Path.GetFileNameWithoutExtension(inputCsvPath);
             fileName += $"_{suffix}.csv";
 
             return Path.Combine(folderPath, fileName);
@@ -146,7 +144,7 @@ namespace N26ToYNAB
         {
             var memoBuilder = new StringBuilder();
 
-            memoBuilder.Append($"{transaction.TransactionType} - {transaction.Category}");
+            memoBuilder.Append($"{transaction.TransactionType} - {transaction.Reference}");
             if (transaction.IsFxTransaction())
             {
                 memoBuilder.Append(