+++ Changelog.md
@@ -4,9 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-/
+## 0.2.0 - Unreleased
+
+### Added
+- Usage information into CLI help output
+- Serilog as logger
+
+### Changed
+- Upgraded to .NET 7.0
+
+### Fixed
+- Updated N26 CSV model to current format
 
 ## 0.1.0 - 2021-09-25
 ### Added
-- Inital release with capability of converting csv files
+- Initial release with capability of converting csv files